requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
openai==1.3.0
sentence-transformers==2.2.2
qdrant-client==1.6.0
firebase-admin==6.2.0
selenium==4.10.0
backports.tarfile==1.2.0